Lab5/requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
